Split databases out of the IPC monolith
  electron/ipc.ts     2908-3266 -> electron/ipc/databases.ts     (63 handlers)
  electron/preload.ts   601-711 -> electron/preload/databases.ts (68 bindings)
  shared/types.ts     8065-8184 -> shared/api/databases.ts       (67 signatures)

Three things had to move deliberately rather than as part of the block, and each is
a case the earlier extractions did not hit:

chatAborters is genuinely shared state, contrary to what the pre-split measurement
concluded. Three separate streaming chats key into the same registry —
worldChat:stream, db:chatStream and research:chatStream — each paired with its own
:cancel channel. Since requestIds are unique per request, giving each domain its own
map would behave identically, but that is a semantic change and does not belong in a
move, so IpcContext carries the registry and documents why.

activeDbChatRequestId sat in preload.ts beside the research-chat equivalent; it moves
into the databases slice, its only user. Likewise dbGuessMime, pendingCsvImports,
CSV_PREVIEW_ROWS and the two BULK_WALK guards, which were module-level in ipc.ts.

electron/ipc.ts is now 4,513 lines (from 6,013 — down 25%), electron/preload.ts
1,428 (from 1,862 — down 23%), shared/types.ts 8,978 (from 9,553). 312 channels
live in a domain module.

test-databases.mjs asserted two channels by matching the text of ipc.ts and
preload.ts; it now asks the census whether they are wired.

Verified: 1,425 tests, lint and typecheck clean, contract test green.
