+++ shared/types.ts
@@ -4,6 +4,7 @@
 // renderer surface stays flat and unchanged; only the file each method is
 // declared in moves. See shared/api/ for the slices extracted so far.
 import type { ProsopographyApi } from './api/prosopography';
+import type { DatabasesApi } from './api/databases';
 import type { TeachingApi } from './api/teaching';
 import type { ToolkitApi } from './api/toolkit';
 import type { TestimoniesApi } from './api/testimonies';
@@ -163,8 +164,6 @@ import type {
 } from './toolkitApps';
 import type {
 } from './protectTypes';
-import type { CsvImportPlanData } from './databaseCsv';
-import type { BulkAttachOptions } from './databaseBulk';
 import type {
   StudyAnnotation,
   StudyAnnotationInput,
@@ -555,20 +554,6 @@ import type { ArchiveMatchMode, ArchiveSortKey } from './archiveFilters';
 export type { ArchiveMatchMode, ArchiveSortKey } from './archiveFilters';
 export type { VaultType };
 import type {
-  DatabaseAttachment,
-  DatabaseColumn,
-  DatabaseColumnConfig,
-  DatabaseColumnType,
-  DatabaseDetail,
-  DatabaseRelation,
-  DatabaseRow,
-  DatabaseRowHit,
-  DatabaseRowSort,
-  DatabaseSearchHit,
-  DatabaseSelectOption,
-  DatabaseSummary,
-  RelationTarget,
-  RelationTargetKind,
 } from './databases';
 export type {
   DatabaseAttachment,
@@ -586,7 +571,6 @@ export type {
   RelationTarget,
   RelationTargetKind,
 } from './databases';
-import type { DatabaseSavedView, SavedViewInput } from './databaseFilters';
 export type {
   DatabaseFilterState,
   DatabaseSavedView,
@@ -595,12 +579,10 @@ export type {
   SavedViewInput,
   SortRule,
 } from './databaseFilters';
-import type { DatabaseProfile } from './dataProfile';
 export type { ColumnProfile, DatabaseProfile, DistributionSlice, NumberStats } from './dataProfile';
 import type { DbChatTurn } from './databaseChat';
 export type { DbChatTurn } from './databaseChat';
 export type { ChartSpec, ChatSegment } from './chartSpec';
-import type { AnalysisRequest, AnalysisResult, AnalysisSuggestion } from './analysisSpec';
 export type {
   AnalysisKind,
   AnalysisOptions,
@@ -7313,7 +7295,7 @@ export interface TestimonyExportResult {
 // IPC API surface exposed on window.nodus via the preload bridge.
 // ─────────────────────────────────────────────────────────────────────────────
 
-export interface NodusApi extends ProsopographyApi, TestimoniesApi, ToolkitApi, TeachingApi {
+export interface NodusApi extends ProsopographyApi, TestimoniesApi, ToolkitApi, TeachingApi, DatabasesApi {
   // settings + secrets
   getSettings(): Promise<AppSettings>;
   updateSettings(patch: Partial<AppSettings>): Promise<AppSettings>;
@@ -8062,126 +8044,6 @@ export interface NodusApi extends ProsopographyApi, TestimoniesApi, ToolkitApi, 
   restorePrimarySourceResearchPackage(
     name?: string | null
   ): Promise<import('./primarySourcesTypes').PrimarySourceRestoreReport | null>;
-  // databases mode (Notion-like structured data)
-  listDatabases(): Promise<DatabaseSummary[]>;
-  searchDatabases(query: string, includeContent: boolean): Promise<DatabaseSearchHit[]>;
-  searchDatabaseRows(query: string, limit?: number): Promise<DatabaseRowHit[]>;
-  getDatabase(id: string): Promise<DatabaseSummary | null>;
-  getDatabaseDetail(id: string): Promise<DatabaseDetail | null>;
-  databaseStats(id: string): Promise<{ rowCount: number; vaultTotal: number; percent: number }>;
-  createDatabase(name: string, icon?: string | null): Promise<DatabaseSummary>;
-  renameDatabase(id: string, name: string): Promise<DatabaseSummary | null>;
-  setDatabaseIcon(id: string, icon: string | null): Promise<DatabaseSummary | null>;
-  deleteDatabase(id: string): Promise<void>;
-  reorderDatabases(ids: string[]): Promise<void>;
-  createDatabaseColumn(
-    databaseId: string,
-    name: string,
-    type: DatabaseColumnType,
-    config?: DatabaseColumnConfig
-  ): Promise<DatabaseColumn>;
-  updateDatabaseColumn(
-    id: string,
-    patch: { name?: string; type?: DatabaseColumnType; config?: DatabaseColumnConfig }
-  ): Promise<DatabaseColumn | null>;
-  deleteDatabaseColumn(id: string): Promise<void>;
-  reorderDatabaseColumns(databaseId: string, ids: string[]): Promise<void>;
-  addDatabaseOption(columnId: string, label: string, color?: string | null): Promise<DatabaseSelectOption>;
-  updateDatabaseOption(id: string, patch: { label?: string; color?: string | null }): Promise<void>;
-  deleteDatabaseOption(id: string): Promise<void>;
-  reorderDatabaseOptions(columnId: string, ids: string[]): Promise<void>;
-  listDatabaseRows(
-    databaseId: string,
-    opts?: { sort?: DatabaseRowSort; limit?: number; offset?: number }
-  ): Promise<DatabaseRow[]>;
-  getDatabaseRow(id: string): Promise<DatabaseRow | null>;
-  createDatabaseRow(databaseId: string): Promise<DatabaseRow>;
-  deleteDatabaseRow(id: string): Promise<void>;
-  setDatabaseCell(rowId: string, columnId: string, raw: string | null): Promise<DatabaseRow | null>;
-  runDatabaseComparisonCell(rowId: string, columnId: string): Promise<DatabaseRow | null>;
-  runDatabaseComparisonColumn(databaseId: string, columnId: string): Promise<{ done: number }>;
-  onDatabaseComparisonProgress(
-    cb: (payload: { vaultId: string; databaseId: string; columnId: string; done: number; total: number }) => void
-  ): () => void;
-  listDatabaseAttachments(rowId: string, columnId: string): Promise<DatabaseAttachment[]>;
-  getDatabaseAttachmentBlob(id: string): Promise<Uint8Array | null>;
-  /**
-   * The attachment's downscaled preview, falling back to the original when there is none.
-   * `mimeType` describes the returned bytes, which is not the attachment's own type when a
-   * thumb was generated.
-   */
-  getDatabaseAttachmentThumb(id: string): Promise<{ bytes: Uint8Array; mimeType: string | null } | null>;
-  deleteDatabaseAttachment(id: string): Promise<void>;
-  downloadDatabaseAttachment(id: string): Promise<{ canceled: boolean; path: string | null }>;
-  pickAndAttachDatabaseFiles(
-    rowId: string,
-    columnId: string
-  ): Promise<{ added: number; attachments: DatabaseAttachment[] }>;
-  runDatabaseAiCell(rowId: string, columnId: string): Promise<string>;
-  runDatabaseAiColumn(databaseId: string, columnId: string): Promise<{ done: number; failed: number }>;
-  generateDatabaseAiImage(rowId: string, columnId: string): Promise<DatabaseAttachment>;
-  generateDatabaseAiImageColumn(databaseId: string, columnId: string): Promise<{ done: number; failed: number }>;
-  onDatabaseAiProgress(
-    cb: (payload: { vaultId: string; databaseId: string; columnId: string; done: number; total: number }) => void
-  ): () => void;
-  listDatabaseRelations(rowId: string, columnId: string): Promise<DatabaseRelation[]>;
-  addDatabaseRelation(
-    rowId: string,
-    columnId: string,
-    targetKind: RelationTargetKind,
-    targetId: string,
-    targetVaultId?: string | null
-  ): Promise<DatabaseRelation>;
-  removeDatabaseRelation(id: string): Promise<void>;
-  searchDatabaseRelationTargets(
-    kind: RelationTargetKind,
-    query: string,
-    databaseId?: string
-  ): Promise<RelationTarget[]>;
-  parseCsvForImport(): Promise<CsvImportPlanData | null>;
-  createDatabaseFromCsv(
-    name: string,
-    headers: string[],
-    rows: string[][],
-    types: (DatabaseColumnType | null)[]
-  ): Promise<DatabaseSummary>;
-  /** Import the CSV held in the main process behind `token`. A null type discards the column. */
-  createDatabaseFromCsvToken(
-    token: string,
-    name: string,
-    types: (DatabaseColumnType | null)[]
-  ): Promise<DatabaseSummary>;
-  /** Drop a parsed CSV the user decided not to import, so its rows stop occupying memory. */
-  releaseCsvImport(token: string): Promise<void>;
-  onCsvImportProgress(cb: (p: { done: number; total: number; finished: boolean }) => void): () => void;
-  exportDatabase(databaseId: string, format: 'csv' | 'json' | 'xlsx'): Promise<{ canceled: boolean; path?: string }>;
-  getDatabaseProfile(databaseId: string): Promise<{ databaseName: string; profile: DatabaseProfile } | null>;
-  analyzeDatabaseReport(databaseId: string): Promise<{ databaseName: string; profileText: string; report: string }>;
-  suggestDatabaseAnalyses(databaseId: string): Promise<{ databaseName: string; suggestions: AnalysisSuggestion[] }>;
-  runDatabaseAnalysis(databaseId: string, request: AnalysisRequest): Promise<{ databaseName: string; request: AnalysisRequest; result: AnalysisResult }>;
-  narrateDatabaseAnalysis(result: AnalysisResult): Promise<string>;
-  dbChatStream(request: DatabaseChatRequest, handlers: { onDelta: (delta: string) => void }): Promise<{ text: string }>;
-  cancelDbChat(): Promise<void>;
-  listDatabaseChatConversations(): Promise<DatabaseChatConversationSummary[]>;
-  getDatabaseChatConversation(id: string): Promise<DatabaseChatConversation | null>;
-  createDatabaseChatConversation(input: { title: string; databaseIds: string[] }): Promise<DatabaseChatConversation>;
-  saveDatabaseChatConversation(id: string, messages: DbChatTurn[], databaseIds: string[]): Promise<DatabaseChatConversation | null>;
-  deleteDatabaseChatConversation(id: string): Promise<void>;
-  listDatabaseViews(databaseId: string): Promise<DatabaseSavedView[]>;
-  createDatabaseView(databaseId: string, input: SavedViewInput): Promise<DatabaseSavedView>;
-  updateDatabaseView(id: string, patch: Partial<SavedViewInput>): Promise<DatabaseSavedView | null>;
-  deleteDatabaseView(id: string): Promise<void>;
-  pickBulkDatabaseFiles(mode?: 'files' | 'folder'): Promise<{ name: string; path: string }[]>;
-  bulkAttachDatabaseFiles(
-    databaseId: string,
-    refColumnId: string,
-    attachmentColumnId: string,
-    files: { name: string; path: string }[],
-    options?: BulkAttachOptions
-  ): Promise<{ attached: number; matched: number; unmatched: number }>;
-  onDatabaseBulkProgress(
-    cb: (payload: { databaseId: string; done: number; total: number; attached: number; matched: number; finished: boolean }) => void
-  ): () => void;
   getMcpStatus(): Promise<McpServerStatus>;
   regenerateMcpToken(): Promise<string>;
   getMcpTunnelStatus(): Promise<McpTunnelStatus>;
